+++ packages/web/src/app/[lng]/memos/setting/component/Setting/index.tsx
@@ -1,109 +1,16 @@
 'use client';
 
-import { MOTION_VARIANTS, PATHS, QUERY_KEY } from '@extension/shared/constants';
-import {
-  useCategoryDeleteMutation,
-  useCategoryPostMutation,
-  useCategoryQuery,
-  useCategoryUpsertMutation,
-} from '@extension/shared/hooks';
-import { LocalStorage } from '@extension/shared/modules/local-storage';
-import { Button } from '@src/components/ui/button';
-import { Label } from '@src/components/ui/label';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@src/components/ui/select';
-import { useLanguage, useSupabaseClient } from '@src/hooks';
-import { useToast } from '@src/hooks/use-toast';
+import { MOTION_VARIANTS } from '@extension/shared/constants';
 import { LanguageType } from '@src/modules/i18n';
-import useTranslation from '@src/modules/i18n/client';
-import { useQueryClient } from '@tanstack/react-query';
 import { motion } from 'framer-motion';
-import { TrashIcon } from 'lucide-react';
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-import { useForm } from 'react-hook-form';
+
+import SettingCategoryForm from './SettingCategoryForm';
+import SettingGuide from './SettingGuide';
+import SettingLanguage from './SettingLanguage';
 
 interface SettingProps extends LanguageType {}
 
-interface CategoryForm {
-  categories: {
-    id: number;
-    name: string;
-  }[];
-}
-
 export default function Setting({ lng }: SettingProps) {
-  const { t } = useTranslation(lng);
-  const { language, setLanguageRouter } = useLanguage();
-  const router = useRouter();
-  const supabaseClient = useSupabaseClient();
-  const { categories } = useCategoryQuery({ supabaseClient });
-  const { mutate: deleteCategory } = useCategoryDeleteMutation({ supabaseClient });
-  const { mutate: upsertCategory } = useCategoryUpsertMutation({ supabaseClient });
-  const { mutate: insertCategory } = useCategoryPostMutation({ supabaseClient });
-  const { register, handleSubmit, setValue } = useForm<CategoryForm>({
-    defaultValues: {
-      categories: [],
-    },
-  });
-  const { toast } = useToast();
-  const queryClient = useQueryClient();
-
-  const handleRestartGuide = () => {
-    LocalStorage.remove('guide');
-    router.push(PATHS.memos);
-  };
-
-  const handleAddCategory = () => {
-    insertCategory(
-      { name: '' },
-      {
-        onSuccess: () => {
-          queryClient.invalidateQueries({ queryKey: QUERY_KEY.category() });
-          toast({
-            title: t('toastTitle.addSuccess'),
-          });
-        },
-      },
-    );
-  };
-
-  const handleCategoryDelete = (id: number) => {
-    deleteCategory(
-      { id },
-      {
-        onSuccess: () => {
-          queryClient.invalidateQueries({ queryKey: QUERY_KEY.category() });
-          toast({
-            title: t('toastTitle.deleteSuccess'),
-          });
-        },
-      },
-    );
-  };
-
-  const onCategoryFormSubmit = (data: CategoryForm) => {
-    upsertCategory(
-      { categoryRequest: data.categories },
-      {
-        onSuccess: () => {
-          queryClient.invalidateQueries({ queryKey: QUERY_KEY.category() });
-          toast({
-            title: t('toastTitleㅋ.saveSuccess'),
-          });
-        },
-      },
-    );
-  };
-
-  useEffect(() => {
-    if (!categories) return;
-
-    setValue(
-      'categories',
-      categories.map(category => ({ id: category.id, name: category.name })),
-    );
-  }, [categories, setValue]);
-
   return (
     <motion.section
       className="grid gap-6"
@@ -111,54 +18,9 @@ export default function Setting({ lng }: SettingProps) {
       initial="initial"
       animate="animate"
       exit="exit">
-      <div className="grid grid-cols-12">
-        <Label className="col-span-4 grid place-items-center">{t('setting.language')}</Label>
-        <Select onValueChange={setLanguageRouter} value={language} aria-label={t('setting.selectLanguage')}>
-          <SelectTrigger className="w-[180px]">
-            <SelectValue placeholder="Theme" />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectItem value="ko">한글</SelectItem>
-            <SelectItem value="en">English</SelectItem>
-          </SelectContent>
-        </Select>
-      </div>
-      <div className="grid grid-cols-12">
-        <Label className="col-span-4 grid place-items-center">{t('setting.guide')}</Label>
-        <Button variant="outline" className="col-span-2" onClick={handleRestartGuide}>
-          {t('setting.restartGuide')}
-        </Button>
-      </div>
-      <form className="grid grid-cols-12" onSubmit={handleSubmit(onCategoryFormSubmit)}>
-        <Label className="col-span-4 grid place-items-center">{t('setting.category')}</Label>
-        <div className="col-span-6 space-y-2">
-          {categories?.map(({ id }, index) => (
-            <div key={id} className="flex items-center gap-2">
-              <input
-                type="text"
-                {...register(`categories.${index}.name`)}
-                className="border-input focus-visible:ring-ring h-9 w-full rounded-md border bg-transparent px-3 py-1 text-sm shadow-sm transition-colors focus-visible:outline-none focus-visible:ring-1"
-              />
-              <Button
-                variant="ghost"
-                size="icon"
-                className="text-destructive h-9 w-9"
-                type="button"
-                onClick={() => handleCategoryDelete(id)}>
-                <TrashIcon size={16} />
-              </Button>
-            </div>
-          ))}
-          <Button variant="outline" className="w-full" onClick={handleAddCategory} type="button">
-            {t('setting.addCategory')}
-          </Button>
-        </div>
-        <div className="col-span-2 grid place-items-center">
-          <Button type="submit" variant="outline">
-            {t('common.save')}
-          </Button>
-        </div>
-      </form>
+      <SettingLanguage lng={lng} />
+      <SettingGuide lng={lng} />
+      <SettingCategoryForm lng={lng} />
     </motion.section>
   );
 }
